requirements.txt
alabaster==0.7.9
alembic==0.8.7
amqp==2.0.3
anyjson==0.3.3
apipkg==1.4
appnope==0.1.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.0
blinker==1.4
celery==3.1.23
cffi==1.7.0
click==6.7
coverage==4.2
cryptography==1.4
decorator==4.0.10
docutils==0.12
execnet==1.4.1
flake8==3.2.1
Flask==0.12
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.0
-e git+https://github.com/ics/Flask-GnuPG#egg=Flask-GnuPG
Flask-JsonSchema==0.1.1
flask-ldap3-login==0.9.10
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
-e git+https://github.com/certeu/Flask-Tinyclients#egg=Flask-Tinyclients
Flask-WTF==0.12
future==0.15.2
gnureadline==6.3.3
httplib2==0.9.2
idna==2.1
imagesize==0.7.1
ipython==5.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.4
jsonschema==2.5.1
kombu==3.0.35
ldap3==1.4.0
lxml==3.6.1
mailmanclient==1.0.1
Mako==1.0.4
MarkupSafe==0.23
mccabe==0.5.3
ndg-httpsclient==0.4.2
onetimepass==1.0.1
passlib==1.6.5
path.py==8.2.1
pbr==1.10.0
pefile==2016.3.28
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.4
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pyClamd==0.3.17
pycodestyle==2.2.0
pycparser==2.14
pyflakes==1.3.0
Pygments==2.1.3
PyMySQL==0.7.6
pyOpenSSL==16.0.0
PyQRCode==1.2.1
pyrabbit==1.1.0
pytest==2.9.2
pytest-cov==2.3.0
pytest-flake8==0.6
pytest-xdist==1.14
python-editor==1.0.1
python-gnupg==0.3.8
python-magic==0.4.12
pytz==2016.6.1
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.4
sphinx-rtd-theme==0.1.9
sphinxcontrib-httpdomain==1.5.0
SQLAlchemy==1.0.14
ssdeep==3.1.1
traitlets==4.2.2
vine==1.1.1
wcwidth==0.1.7
Werkzeug==0.11.15
WTForms==2.1
